requirements.txt
speedtest-cli==2.1.3  # 2.1.3のバージョンを使用する
ping3
schedule
pandas
matplotlib